refactor: split monolithic canvas.rs into composable modules
Backend: decompose 2164-line apps/papillon/src/commands/canvas.rs into
14 focused sub-modules under commands/canvas/:
  types, outcome, intent, resolution, execution, helpers,
  orchestration, approval, crud, blocks, messages, notes, guide, mod

mod.rs acts as a re-export facade — all existing lib.rs generate_handler![]
paths (commands::canvas::*) remain unchanged. 134 tests pass.

Frontend: decompose 954-line pages/canvas.rs into 7 extracted components
under frontend/src/components/:
  canvas_back_face, canvas_chat_thread, canvas_empty_state,
  canvas_workflow_pipeline, hitl_gate, outcome_summary, pipeline_builder_tab

pages/canvas.rs reduced to 125 lines (root CanvasPage composition only).
cargo check --target wasm32-unknown-unknown passes with zero errors.
