+++ apps/papillon/frontend/src/components/outcome_summary.rs
@@ -1,0 +1,217 @@
+use leptos::prelude::*;
+use serde::{Deserialize, Serialize};
+use wasm_bindgen_futures::spawn_local;
+
+use crate::bridge;
+
+/// A synthesised summary of a single completed agent interaction episode.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct EpisodeSummary {
+    pub session_did: String,
+    pub agent_did: String,
+    pub agent_name: String,
+    pub action: String,
+    pub outcome: String,
+    pub timestamp: String,
+    pub receipt_hash: String,
+    pub intent_summary: Option<String>,
+}
+
+/// Synthesised canvas state returned by the `get_canvas_state` Tauri command.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CanvasSummaryState {
+    pub principal_did: String,
+    pub episodes: Vec<EpisodeSummary>,
+    pub success_count: u32,
+    pub failure_count: u32,
+    pub active_sessions: u32,
+}
+
+/// Outcome synthesis panel — loads completed episode summaries from the
+/// persistent episode store via the `get_canvas_state` Tauri command and
+/// renders them as a timeline with wing-spectrum status indicators.
+///
+/// - Teal  (#2ec4a0) — success / resolved
+/// - Gold  (#f0a030) — in-progress (active sessions)
+/// - Coral (#e8706a) — failure / error
+///
+/// Only rendered when Tauri IPC is available; silently hidden in browser mode.
+#[component]
+pub fn OutcomeSummary() -> impl IntoView {
+    let canvas_data = RwSignal::new(None::<CanvasSummaryState>);
+    let loading = RwSignal::new(false);
+    let error = RwSignal::new(None::<String>);
+
+    // Load canvas state on mount (Tauri path only)
+    Effect::new(move || {
+        if !bridge::tauri_available() {
+            return;
+        }
+        loading.set(true);
+        spawn_local(async move {
+            match bridge::invoke_no_args::<CanvasSummaryState>("get_canvas_state").await {
+                Ok(state) => {
+                    canvas_data.set(Some(state));
+                    error.set(None);
+                }
+                Err(e) => {
+                    error.set(Some(e));
+                }
+            }
+            loading.set(false);
+        });
+    });
+
+    let has_episodes = move || {
+        canvas_data
+            .get()
+            .as_ref()
+            .map(|s| !s.episodes.is_empty())
+            .unwrap_or(false)
+    };
+
+    view! {
+        <Show when=move || bridge::tauri_available()>
+            <div class="outcome-summary">
+                <div class="outcome-summary-header">
+                    <span class="outcome-summary-label">"OUTCOME_SYNTHESIS"</span>
+                    <Show when=move || loading.get()>
+                        <span class="outcome-loading-indicator">"..."</span>
+                    </Show>
+                </div>
+
+                <Show when=move || error.get().is_some()>
+                    <div class="outcome-error">
+                        {move || error.get().unwrap_or_default()}
+                    </div>
+                </Show>
+
+                <Show when=move || canvas_data.get().is_some() && !loading.get()>
+                    {move || {
+                        let state = canvas_data.get().unwrap();
+                        let did_short = if state.principal_did.len() > 20 {
+                            format!("{}...", &state.principal_did[..20])
+                        } else {
+                            state.principal_did.clone()
+                        };
+                        view! {
+                            <div class="outcome-principal">
+                                <span class="outcome-key">"DID"</span>
+                                <span class="outcome-did">{did_short}</span>
+                            </div>
+                            <div class="outcome-stats">
+                                <div class="outcome-stat outcome-stat-success">
+                                    <span class="outcome-stat-icon">
+                                        // Checkmark SVG — success indicator
+                                        <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5">
+                                            <path d="M20 6 9 17l-5-5"/>
+                                        </svg>
+                                    </span>
+                                    <span class="outcome-stat-val">{state.success_count}</span>
+                                </div>
+                                <div class="outcome-stat outcome-stat-failure">
+                                    <span class="outcome-stat-icon">
+                                        // X SVG — failure indicator
+                                        <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5">
+                                            <path d="M18 6 6 18M6 6l12 12"/>
+                                        </svg>
+                                    </span>
+                                    <span class="outcome-stat-val">{state.failure_count}</span>
+                                </div>
+                                <div class="outcome-stat outcome-stat-active">
+                                    <span class="outcome-stat-icon">
+                                        // Clock SVG — in-progress indicator
+                                        <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5">
+                                            <circle cx="12" cy="12" r="10"/>
+                                            <path d="M12 6v6l4 2"/>
+                                        </svg>
+                                    </span>
+                                    <span class="outcome-stat-val">{state.active_sessions}</span>
+                                </div>
+                            </div>
+                        }
+                    }}
+                </Show>
+
+                <Show when=has_episodes>
+                    <div class="outcome-divider" />
+                    <div class="outcome-timeline">
+                        <For
+                            each=move || {
+                                canvas_data
+                                    .get()
+                                    .map(|s| s.episodes)
+                                    .unwrap_or_default()
+                            }
+                            key=|ep| ep.receipt_hash.clone()
+                            children=move |ep| {
+                                let action_display = ep
+                                    .action
+                                    .strip_prefix("schema:")
+                                    .unwrap_or(&ep.action)
+                                    .to_string();
+                                let ts_short = ep.timestamp.chars().take(16).collect::<String>();
+                                let outcome_class = match ep.outcome.as_str() {
+                                    "success" => "outcome-entry outcome-entry-success",
+                                    "failure" | "rejected" => "outcome-entry outcome-entry-failure",
+                                    _ => "outcome-entry outcome-entry-pending",
+                                };
+                                // Outcome icon rendered as an inline SVG via a
+                                // type-erased AnyView to avoid match-arm type mismatch.
+                                let outcome_icon = match ep.outcome.as_str() {
+                                    "success" => view! {
+                                        <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5">
+                                            <path d="M20 6 9 17l-5-5"/>
+                                        </svg>
+                                    }.into_any(),
+                                    "failure" | "rejected" => view! {
+                                        <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5">
+                                            <path d="M18 6 6 18M6 6l12 12"/>
+                                        </svg>
+                                    }.into_any(),
+                                    _ => view! {
+                                        <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2.5">
+                                            <circle cx="12" cy="12" r="10"/>
+                                            <path d="M12 6v6l4 2"/>
+                                        </svg>
+                                    }.into_any(),
+                                };
+                                let agent_name = ep.agent_name.clone();
+                                let intent = ep.intent_summary.clone();
+                                view! {
+                                    <div class=outcome_class>
+                                        <span class="outcome-entry-icon">{outcome_icon}</span>
+                                        <div class="outcome-entry-body">
+                                            <div class="outcome-entry-top">
+                                                <span class="outcome-entry-agent">{agent_name}</span>
+                                                <span class="outcome-entry-action">{action_display}</span>
+                                            </div>
+                                            {move || {
+                                                if let Some(ref summary) = intent {
+                                                    let s = summary.clone();
+                                                    view! {
+                                                        <div class="outcome-entry-intent">{s}</div>
+                                                    }.into_any()
+                                                } else {
+                                                    view! { <span /> }.into_any()
+                                                }
+                                            }}
+                                            <div class="outcome-entry-meta">
+                                                <span class="outcome-entry-ts">{ts_short}</span>
+                                                <span class="outcome-entry-hash">{ep.receipt_hash.clone()}</span>
+                                            </div>
+                                        </div>
+                                    </div>
+                                }
+                            }
+                        />
+                    </div>
+                </Show>
+
+                <Show when=move || !has_episodes() && !loading.get() && bridge::tauri_available()>
+                    <div class="outcome-empty">"NO_COMPLETED_EPISODES"</div>
+                </Show>
+            </div>
+        </Show>
+    }
+}
