+++ apps/papillon/src/commands/canvas/mod.rs
@@ -1,0 +1,278 @@
+#![allow(clippy::unwrap_used)]
+
+mod approval;
+mod blocks;
+mod crud;
+mod execution;
+mod guide;
+mod helpers;
+mod intent;
+mod messages;
+mod notes;
+mod orchestration;
+mod outcome;
+mod resolution;
+mod types;
+
+// ── Public Tauri commands (referenced by lib.rs generate_handler![]) ──────
+// Use glob re-exports so `__cmd__<fn>` symbols generated by #[tauri::command]
+// are also visible at the `commands::canvas::*` path used by generate_handler![].
+pub use approval::*;
+pub use blocks::*;
+pub use crud::*;
+pub use guide::*;
+pub use messages::*;
+pub use notes::*;
+pub use orchestration::*;
+pub use outcome::*;
+
+// ── Internal re-exports used by sibling command modules ──────────────────
+pub(crate) use resolution::resolve_agent;
+
+// ── Internal constants ────────────────────────────────────────────────────
+/// BM25 confidence threshold for Level 2 intent routing.
+/// Queries below this threshold fall through to the Level 3 on-device / federation NLU path.
+pub(crate) const INTENT_CONFIDENCE_THRESHOLD: f32 = 0.25;
+
+// ── General unit tests ────────────────────────────────────────────────────
+
+#[cfg(test)]
+mod tests {
+    use super::execution::assess_handshake_quality;
+    use super::intent::map_label_to_action;
+    use crate::handshake;
+
+    // ── map_label_to_action ────────────────────────────────────
+
+    #[test]
+    fn map_label_to_action_coverage() {
+        // All 14 explicit labels map to non-empty action + agent names.
+        let labels = [
+            "weather",
+            "currency-exchange",
+            "dictionary",
+            "code-repository",
+            "book",
+            "tech-news",
+            "academic-paper",
+            "geocode",
+            "dataset",
+            "music",
+            "film",
+            "job-listing",
+            "recipe",
+            "product",
+        ];
+        for label in &labels {
+            let (action, agent) = map_label_to_action(label);
+            assert!(!action.is_empty(), "action empty for label: {label}");
+            assert!(!agent.is_empty(), "agent empty for label: {label}");
+            assert!(
+                action.starts_with("schema:"),
+                "action not schema: prefixed for label: {label}"
+            );
+        }
+    }
+
+    #[test]
+    fn map_label_to_action_unknown_falls_back_to_search() {
+        let (action, agent) = map_label_to_action("completely-unknown-intent");
+        assert_eq!(action, "schema:SearchAction");
+        assert_eq!(agent, "DuckDuckGo Search");
+    }
+
+    #[test]
+    fn map_label_to_action_question_answer_not_mapped() {
+        // "question-answer" is handled by the confidence check in classify_intent
+        // and never reaches map_label_to_action. But if it did, it falls back.
+        let (action, _) = map_label_to_action("question-answer");
+        assert_eq!(action, "schema:SearchAction");
+    }
+
+    #[test]
+    fn score_agent_prefers_keyword_match_without_profile() {
+        // No DB available in unit tests, so we verify the scoring constants
+        // used in score_agent: preferred baseline (0.6) > non-preferred (0.4)
+        let preferred_baseline = 0.6_f64;
+        let non_preferred_baseline = 0.4_f64;
+        assert!(preferred_baseline > non_preferred_baseline);
+    }
+
+    #[test]
+    fn assess_quality_empty_result() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"agent": "test"}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        // No "result" key → 0.3
+        assert!((quality - 0.3).abs() < f64::EPSILON);
+    }
+
+    #[test]
+    fn assess_quality_null_result() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": null}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.2).abs() < f64::EPSILON);
+    }
+
+    #[test]
+    fn assess_quality_rich_object() {
+        let result = handshake::HandshakeResult {
+            schema_type: "SearchResult".into(),
+            content: serde_json::json!({
+                "result": {
+                    "title": "Test",
+                    "url": "https://example.com",
+                    "snippet": "A test result",
+                    "source": "DuckDuckGo"
+                }
+            }),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        // 4 fields → 0.8
+        assert!((quality - 0.8).abs() < f64::EPSILON);
+    }
+
+    #[test]
+    fn assess_quality_large_array() {
+        let result = handshake::HandshakeResult {
+            schema_type: "SearchResult".into(),
+            content: serde_json::json!({
+                "result": ["a", "b", "c", "d"]
+            }),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        // Array with 4 items → 0.8 (generic match)
+        assert!(quality >= 0.8);
+    }
+
+    #[test]
+    fn assess_quality_empty_string() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": ""}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.3).abs() < f64::EPSILON);
+    }
+
+    // ── assess_handshake_quality — uncovered branches ─────────
+
+    #[test]
+    fn assess_quality_short_string_scores_medium() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": "short answer"}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.5).abs() < f64::EPSILON, "got {quality}");
+    }
+
+    #[test]
+    fn assess_quality_long_string_scores_high() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": "this is a longer answer that exceeds twenty characters"}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.8).abs() < f64::EPSILON, "got {quality}");
+    }
+
+    #[test]
+    fn assess_quality_one_field_object_scores_medium() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": {"name": "Rust"}}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.6).abs() < f64::EPSILON, "got {quality}");
+    }
+
+    #[test]
+    fn assess_quality_six_field_object_scores_perfect() {
+        let result = handshake::HandshakeResult {
+            schema_type: "SearchResult".into(),
+            content: serde_json::json!({
+                "result": {
+                    "a": 1, "b": 2, "c": 3,
+                    "d": 4, "e": 5, "f": 6
+                }
+            }),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 1.0).abs() < f64::EPSILON, "got {quality}");
+    }
+
+    #[test]
+    fn assess_quality_small_array_scores_medium() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": ["only_one"]}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.6).abs() < f64::EPSILON, "got {quality}");
+    }
+
+    #[test]
+    fn assess_quality_empty_array_scores_low() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": []}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.3).abs() < f64::EPSILON, "got {quality}");
+    }
+
+    #[test]
+    fn assess_quality_boolean_result_scores_high() {
+        let result = handshake::HandshakeResult {
+            schema_type: "Thing".into(),
+            content: serde_json::json!({"result": true}),
+            agent_name: "Test".into(),
+        };
+        let quality = assess_handshake_quality(&result);
+        assert!((quality - 0.8).abs() < f64::EPSILON, "got {quality}");
+    }
+
+    // ── score_agent — constant validation ─────────────────────
+
+    #[test]
+    fn score_agent_cold_start_preferred_beats_non_preferred() {
+        let preferred_cold: f64 = 0.6;
+        let non_preferred_cold: f64 = 0.4;
+        assert!(
+            preferred_cold > non_preferred_cold,
+            "cold-start preferred score ({preferred_cold}) must beat non-preferred ({non_preferred_cold})"
+        );
+        assert!((0.0..=1.0).contains(&preferred_cold));
+        assert!((0.0..=1.0).contains(&non_preferred_cold));
+    }
+
+    #[test]
+    fn score_agent_constants_are_valid_probability_weights() {
+        let w_success: f64 = 0.35;
+        let w_quality: f64 = 0.35;
+        let w_pref: f64 = 0.20;
+        let w_keyword: f64 = 0.10;
+        let sum = w_success + w_quality + w_pref + w_keyword;
+        assert!(
+            (sum - 1.0).abs() < f64::EPSILON,
+            "score_agent weights must sum to 1.0, got {sum}"
+        );
+    }
+}
